บันทึกอุณหภูมิลงฐานข้อมูลและแสดงในหน้า TABLE

+++ app/src/main/java/com/siamatic/tms/database/database.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "temp")
 data class Temp(
     @PrimaryKey(autoGenerate = true) val id: Int = 0, // ID of temperature autogenerate
-    @ColumnInfo(defaultValue = "0.0") val temp1: Double?, // Probe 1 temperature
-    @ColumnInfo(defaultValue = "0.0") val temp2: Double?, // Probe 2 temperature
+    @ColumnInfo(defaultValue = "0.0") val temp1: Float?, // Probe 1 temperature
+    @ColumnInfo(defaultValue = "0.0") val temp2: Float?, // Probe 2 temperature
     val createdAt: Long // timestamp of record added!
 )
